NotificationListener: Send useful for id onDeleteNotification()
Currently the id sent during an onDeleteNotification() is different
to the id supplied (as part of the NotificationSpec) during
onNotification(). There doesn't seem to be any way for the device
coordinator to discover the id it will be sent during a delete since
the relationship between the two id codes is stored in a private
variable in the listener.

Use the lookup table to send the device coordinator a more useful id.