wasp-os: Introduce a new device type
wasp-os is an open-source firmware, based on MicroPython, for smart
watches that are based on the nRF52 family of microcontrollers.
Currently this includes the hacker friendly PineTime from Pine64 as
well as the Colmi P8, which is a popular device with watch modders.

For GadgetBridge integration wasp-os is very similar to handling
Espruino/BangleJS: the watch will interact with the bridge by
providing REPL-over-BLE via the Nordic UART service.

wasp-os implements a similar protocol by using JSON wrapped in a
GB() function call as the primary way to interact. However
there are differences in capability between BangleJS and wasp-os
which leads us to introduce a different device type to handle
wasp-os.